Added BreedListViewModel Logic for better testability
